renamed a library module in web
