+++ web/tomato/templatetags/tomato.py
@@ -5,7 +5,7 @@ from django import template
 from ..lib import getVersion, serverInfo, security_token
 from django.utils.safestring import mark_safe
 from ..lib import anyjson as json
-from ..lib.reference_library import tech_to_label as lib_tech_to_label
+from ..lib.references_web import tech_to_label as lib_tech_to_label
 
 from ..lib.settings import get_settings
 from .. import settings as config_module
